requirements.txt
pyahocorasick==1.3.0
pandas==0.23.4
matplotlib==2.2.3
scikit-learn==0.20.2

#conda activate dnaclass

####For keras:
#conda install mkl 
pip install parameterized #optional but may be required for unit testing.
conda install theano pygpu
#neural networks:
conda install pytorch torchvision cudatoolkit=10.0 -c pytorch
conda install jupyter
pip install kaggle 
